Many Features Complete, Nothing Tested
Started work on the basicGradeCalculator() which is almost complete. It does not save anything to file between sessions but has most of the features which should be present in the final version of this project.

the goal was to get something built and complete that works for my more basic purposes